myapp/backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
pandas==2.1.3
